refactor: irrational num class
